speaker_verification/ecapa_tdnn/requirements.txt
wandb
yacs
tqdm
soundfile
torchinfo
sklearn==0.0
timm
einops
scipy
torchaudio
librosa
